pyproject.toml
[project]
name = "sea-cargo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=4.2.1",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.1",
    "streamlit>=1.42.2",
]
